fix(onboarding): protect the rollout transition
