+++ apps/www/lib/routing/public/ownership.test.ts
@@ -35,6 +35,8 @@ describe("public route ownership", () => {
     ["de", "search"],
     ["de", "pricing"],
     ["de", "onboarding"],
+    ["de", "onboarding/focus"],
+    ["de", "onboarding/role"],
     ["de", "chat/new"],
     ["de", "user/settings/subscriptions"],
     ["de", "user/nabil/chat"],
@@ -64,8 +66,6 @@ describe("public route ownership", () => {
     ["de", "fabricated"],
     ["de", "auth/fabricated"],
     ["de", "chat/id/fabricated"],
-    ["de", "onboarding/focus"],
-    ["de", "onboarding/role"],
     ["de", "onboarding/fabricated"],
     ["de", "user"],
     ["de", "user/settings/fabricated"],
